Ref: update file structure
