+++ lib/cell_states.rb
