+++ lib/helper.rb
@@ -1,7 +1,7 @@
-require_relative 'ship'
-require_relative 'cell'
-require_relative 'board'
-require_relative 'turn'
-require_relative 'player'
-require_relative 'game'
-require_relative 'messages'
+require './lib/ship'
+require './lib/cell'
+require './lib/board'
+require './lib/player'
+require './lib/game'
+require './lib/menu'
+require './lib/messages'
