+++ lib/ship.rb
@@ -19,5 +19,4 @@ class Ship
   def hit
     @health -= 1
   end
-
 end
