--- lib2/helper.rb
@@ -1,7 +1,0 @@
-require './lib2/ship'
-require './lib2/cell'
-require './lib2/board'
-require './lib2/player'
-require './lib2/game'
-require './lib2/menu'
-require './lib2/messages'
