+++ old lib/helper.rb
@@ -1,0 +1,7 @@
+require_relative 'ship'
+require_relative 'cell'
+require_relative 'board'
+require_relative 'turn'
+require_relative 'player'
+require_relative 'game'
+require_relative 'messages'
