+++ old lib/ship.rb
@@ -19,4 +19,5 @@ class Ship
   def hit
     @health -= 1
   end
+
 end
