+++ old lib/ship_runner.rb
