+++ old lib/turn.rb
