+++ spec/spec_helper.rb
@@ -2,5 +2,4 @@ require './lib/ship'
 require './lib/cell'
 require './lib/board'
 require './lib/player'
-require './lib/turn'
 require './lib/messages'
