--- spec/spec_helper2.rb
@@ -1,6 +1,0 @@
-require './lib2/ship'
-require './lib2/cell'
-require './lib2/board'
-require './lib2/player'
-require './lib2/turn'
-require './lib2/messages'
